% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Code for the wanted state}

\item{year}{Wanted year in the format yyyy (four digits)}
}
\value{
A plot of the state with dots locating each accident.

Errors are issued for invalid state numbers. A message is issued when there
are no accidents to plot.
}
\description{
Plot the state's map with dots locating the year's accidents.
}
\examples{
\dontrun{
fars_map_state(1, 2015)
}

}
