% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data into list of months}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of wanted years in the format yyyy (four digits)}
}
\value{
A list of data frames, with one entry per year. Each data frame has a
column with the month of each row from the original file and a column year.

If the file for a year is not found, a warning will be issued, and the
corresponding entry in the list is set to NULL.
}
\description{
Read downloaded data from US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System into a list of months available per year.
}
\examples{
\dontrun{
fars_read_years(2013:2016)
}

}
