% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS data number of entries per year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of wanted years in the format yyyy (four digits)}
}
\value{
A data frame, with the number of entries for the matching year
(columns) and month (rows).

If the file for a year is not found, a warning will be issued and there will
be no column for the year.
}
\description{
Read downloaded data from US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System and delivers a summary with number of
entries per year and month.
}
\examples{
\dontrun{
fars_summarize_years(2012:2015)
}

}
