% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Wanted year in the format yyyy (four digits). Will be coerced to
integer.}
}
\value{
A string ready with the filename to be used as input to
\code{fars_read}.
}
\description{
Get the standard filename format for data from US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System, with the
according year.
}
\examples{
accident_2014 <- make_filename(make_filename(2014))

}
